first commit of Dancebot Rework branch. Aiming for original functionality with the same front end, but with a different backend. Currently webserver API is finished. Needs HTML_JS fixed. Then move on towards dancebot API.
