+++ DemobotLegsESP32/DemobotAPI.cpp
@@ -1,0 +1,8 @@
+/**
+ * Author: Matthew Yu
+ * Last Modified: 05/06/20
+ * Project: Dancebot Swarm
+ * File: DancebotAPI.cpp
+ * Description: Dancebot API to drive hardware with using a state structure
+ * Organization: UT IEEE RAS
+ */
