+++ DemobotLegsESP32/DemobotAPI.h
@@ -1,0 +1,18 @@
+/**
+ * Author: Matthew Yu
+ * Last Modified: 05/06/20
+ * Project: Dancebot Swarm
+ * File: DancebotAPI.h
+ * Description: Dancebot API to drive hardware with using a state structure
+ * Organization: UT IEEE RAS
+ */
+#ifndef DEMOBOTAPI
+#define DEMOBOTAPI
+
+#include "DancingServos.h"
+#include "WebController.h"
+
+void initializeRobot();
+
+
+#endif
